Parallel illumina trim

+++ mlst_denovo_velvet_clean.py
@@ -2,7 +2,9 @@
 
 from mlst_modules import rm_files
 import subprocess
+import os
 
 rm_files(['run_mlst_velveth.*', 'run_mlst_velvetg.*', 'run_mlst_interleave.*', '*.interleaved', 'pbsjob.tmp*', 'run_mlst_velvetaccept.*', 'run_mlst_trim.*'])
 
-subprocess.call('rm -r trimmed/', shell=True)
+if os.path.exists('trimmed'):
+   subprocess.call('rm -r trimmed/', shell=True)
